Less verbose logs
Logs are cluttered by 'initial value...updated value' and it is really not
necessary to log that using a 'info' level.
The 'error' log level is excessive for the missing country.

If someone really wants to display the original and updated values, there still
is the possibility to use a ``--log-handler=openerp.addons.base_phone:DEBUG``.