[FIX] voip_oca: filter activities to avoid accessing records the user cannot read
In a multi-company environment, when Odoo searches for activities to display in the systray,
an error is raised:
"Uh-oh! Looks like you have stumbled upon some top-secret records."

This occurs because some activities point to records that the current user does not have
access to, due to multi-company restrictions or other access rules defined in the `res_model`.

Steps to reproduce:

- In a fresh database with demo data, install the `voip_oca` and `account` modules.
- Create a new company.
- Switch to the new company.